pyproject.toml
[project]
name = "productagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.10.0",
    "fastapi>=0.128.0",
    "fastembed>=0.7.4",
    "fastparquet>=2025.12.0",
    "ipykernel>=7.1.0",
    "kaggle>=1.8.3",
    "kagglehub>=0.3.13",
    "langchain>=1.2.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-openai>=1.1.6",
    "pandas>=2.3.3",
    "polars>=1.36.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "redis>=7.1.0",
    "uvicorn>=0.40.0",
]
